Move list to dashboard
