+++ src/panels/config/backup/ha-config-backup-dashboard.ts
@@ -1,17 +1,39 @@
-import "@material/mwc-list/mwc-list";
-import type { TemplateResult } from "lit";
+import { mdiPlus } from "@mdi/js";
+import type { PropertyValues, TemplateResult } from "lit";
 import { css, html, LitElement } from "lit";
-import { customElement, property } from "lit/decorators";
+import { customElement, property, state } from "lit/decorators";
+import memoizeOne from "memoize-one";
+import { relativeTime } from "../../../common/datetime/relative_time";
+import { navigate } from "../../../common/navigate";
+import type { LocalizeFunc } from "../../../common/translations/localize";
+import type {
+  DataTableColumnContainer,
+  RowClickedEvent,
+} from "../../../components/data-table/ha-data-table";
 import "../../../components/ha-button";
 import "../../../components/ha-card";
+import "../../../components/ha-circular-progress";
+import "../../../components/ha-fab";
 import "../../../components/ha-icon";
 import "../../../components/ha-icon-next";
-import "../../../components/ha-list-item";
-import "../../../layouts/hass-subpage";
-
-import { navigate } from "../../../common/navigate";
+import "../../../components/ha-svg-icon";
+import {
+  fetchBackupAgentsSynced,
+  fetchBackupInfo,
+  generateBackup,
+  type BackupContent,
+  type BackupData,
+} from "../../../data/backup";
+import "../../../layouts/hass-tabs-subpage-data-table";
 import { SubscribeMixin } from "../../../mixins/subscribe-mixin";
-import type { HomeAssistant } from "../../../types";
+import type { HomeAssistant, Route } from "../../../types";
+import { brandsUrl } from "../../../util/brands-url";
+import {
+  showAlertDialog,
+  showConfirmationDialog,
+} from "../../lovelace/custom-card-helpers";
+
+const localAgent = "backup.local";
 
 @customElement("ha-config-backup-dashboard")
 class HaConfigBackupDashboard extends SubscribeMixin(LitElement) {
@@ -19,15 +41,107 @@ class HaConfigBackupDashboard extends SubscribeMixin(LitElement) {
 
   @property({ type: Boolean }) public narrow = false;
 
+  @property({ attribute: false }) public route!: Route;
+
+  @state() private _backupData?: BackupData;
+
+  private _columns = memoizeOne(
+    (
+      narrow,
+      _language,
+      localize: LocalizeFunc
+    ): DataTableColumnContainer<BackupContent> => ({
+      name: {
+        title: localize("ui.panel.config.backup.name"),
+        main: true,
+        sortable: true,
+        filterable: true,
+        flex: 2,
+        template: (backup) =>
+          narrow || !backup.path
+            ? backup.name
+            : html`${backup.name}
+                <div class="secondary">${backup.path}</div>`,
+      },
+      path: {
+        title: localize("ui.panel.config.backup.path"),
+        hidden: !narrow,
+        template: (backup) => backup.path || "-",
+      },
+      size: {
+        title: localize("ui.panel.config.backup.size"),
+        filterable: true,
+        sortable: true,
+        template: (backup) => Math.ceil(backup.size * 10) / 10 + " MB",
+      },
+      date: {
+        title: localize("ui.panel.config.backup.created"),
+        direction: "desc",
+        filterable: true,
+        sortable: true,
+        template: (backup) =>
+          relativeTime(new Date(backup.date), this.hass.locale),
+      },
+      locations: {
+        title: "Locations",
+        template: (backup) =>
+          html` ${[
+            ...(backup.path ? [localAgent] : []),
+            ...(backup.agents || []).sort(),
+          ].map((agent) => {
+            const [domain, name] = agent.split(".");
+            return html`
+              <img
+                title=${name}
+                .src=${brandsUrl({
+                  domain,
+                  type: "icon",
+                  useFallback: true,
+                  darkOptimized: this.hass.themes?.darkMode,
+                })}
+                height="24"
+                crossorigin="anonymous"
+                referrerpolicy="no-referrer"
+                alt=${name}
+                slot="graphic"
+              />
+            `;
+          })}`,
+      },
+    })
+  );
+
   protected render(): TemplateResult {
+    const backingUp = true;
+
     return html`
-      <hass-subpage
-        back-path="/config/system"
+      <hass-tabs-subpage-data-table
+        hasFab
+        .tabs=${[
+          {
+            translationKey: "ui.panel.config.backup.caption",
+            path: `/config/backup/list`,
+          },
+        ]}
         .hass=${this.hass}
         .narrow=${this.narrow}
-        .header=${this.hass.localize("ui.panel.config.backup.caption")}
+        back-path="/config/system"
+        clickable
+        id="slug"
+        .route=${this.route}
+        @row-click=${this._showBackupDetails}
+        .columns=${this._columns(
+          this.narrow,
+          this.hass.language,
+          this.hass.localize
+        )}
+        .data=${this._backupData?.backups ?? []}
+        .noDataText=${this.hass.localize("ui.panel.config.backup.no_backups")}
+        .searchLabel=${this.hass.localize(
+          "ui.panel.config.backup.picker.search"
+        )}
       >
-        <div class="header">
+        <div slot="top_header" class="header">
           <ha-card outlined>
             <div class="summary">
               <div class="summary-icon success">
@@ -56,15 +170,84 @@ class HaConfigBackupDashboard extends SubscribeMixin(LitElement) {
                 <p class="summary-description">One is off-site</p>
               </div>
               <div class="summary-action">
-                <ha-button @click=${this._configureBackupLocations}
-                  >Configure</ha-button
-                >
+                <ha-button @click=${this._configureBackupLocations}>
+                  Configure
+                </ha-button>
               </div>
             </div>
           </ha-card>
         </div>
-      </hass-subpage>
+        <ha-fab
+          slot="fab"
+          ?disabled=${backingUp}
+          .label=${backingUp
+            ? this.hass.localize("ui.panel.config.backup.creating_backup")
+            : this.hass.localize("ui.panel.config.backup.create_backup")}
+          extended
+          @click=${this._generateBackup}
+        >
+          ${backingUp
+            ? html`
+                <div slot="icon">
+                  <ha-circular-progress indeterminate></ha-circular-progress>
+                </div>
+              `
+            : html`<ha-svg-icon slot="icon" .path=${mdiPlus}></ha-svg-icon>`}
+        </ha-fab>
+      </hass-tabs-subpage-data-table>
     `;
+  }
+
+  protected firstUpdated(changedProps: PropertyValues) {
+    super.firstUpdated(changedProps);
+    this._fetchBackups();
+  }
+
+  private async _fetchBackups(): Promise<void> {
+    const backupData: Record<string, BackupContent> = {};
+    const [local, synced] = await Promise.all([
+      fetchBackupInfo(this.hass),
+      fetchBackupAgentsSynced(this.hass),
+    ]);
+
+    for (const backup of local.backups) {
+      backupData[backup.slug] = backup;
+    }
+    for (const agent of synced) {
+      if (!(agent.slug in backupData)) {
+        backupData[agent.slug] = { ...agent, agents: [agent.agent_id] };
+      } else if (!("agents" in backupData[agent.slug])) {
+        backupData[agent.slug].agents = [agent.agent_id];
+      } else {
+        backupData[agent.slug].agents!.push(agent.agent_id);
+      }
+    }
+    this._backupData = {
+      backing_up: local.backing_up,
+      backups: Object.values(backupData),
+    };
+  }
+
+  private async _generateBackup(): Promise<void> {
+    const confirm = await showConfirmationDialog(this, {
+      title: this.hass.localize("ui.panel.config.backup.create.title"),
+      text: this.hass.localize("ui.panel.config.backup.create.description"),
+      confirmText: this.hass.localize("ui.panel.config.backup.create.confirm"),
+    });
+    if (!confirm) {
+      return;
+    }
+
+    generateBackup(this.hass)
+      .then(() => this._fetchBackups())
+      .catch((err) => showAlertDialog(this, { text: (err as Error).message }));
+
+    await this._fetchBackups();
+  }
+
+  private _showBackupDetails(ev: CustomEvent): void {
+    const slug = (ev.detail as RowClickedEvent).id;
+    navigate(`/config/backup/details/${slug}`);
   }
 
   private _configureAutomaticBackup() {
@@ -77,10 +260,11 @@ class HaConfigBackupDashboard extends SubscribeMixin(LitElement) {
 
   static styles = css`
     .header {
-      padding: 24px 16px;
+      padding: 16px 16px 0 16px;
       display: flex;
       flex-direction: row;
       gap: 16px;
+      background-color: var(--primary-background-color);
     }
     @media (max-width: 1000px) {
       .header {
@@ -160,6 +344,13 @@ class HaConfigBackupDashboard extends SubscribeMixin(LitElement) {
       text-overflow: ellipsis;
       overflow: hidden;
       white-space: nowrap;
+    }
+    ha-fab[disabled] {
+      --mdc-theme-secondary: var(--disabled-text-color) !important;
+    }
+    ha-circular-progress {
+      --md-sys-color-primary: var(--secondary-text-color);
+      --md-circular-progress-size: 36px;
     }
   `;
 }
